Adding DetailsPro
Hi @jessesquires, I saw you were open to PRs and I think my app DetailsPro is a great resource for designers getting started. 

Full disclosure, DetailsPro does have an optional upgrade but I feel more like I am actually sharing a resource because, as the sole developer of the app, I know that I always intend to keep the app available for free with a fully usable free mode. In fact, planning on expanding it so it'll only get free-er as it grows.  It's pretty fun to just pop into it and play around with stacks and other views to get the hang of designs. 

No worries at all if it's not right for this list

Cheers

p.s. the placement is only a suggestion please feel free to suggest a different heading / grouping